+++ ikii-back-microservices/ikii-back-commons-web/src/main/java/mx/ikii/web/commons/controller/IFileController.java
@@ -1,0 +1,16 @@
+package mx.ikii.web.commons.controller;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.multipart.MultipartFile;
+
+@RestController
+@RequestMapping("/upload")
+public interface IFileController {
+  
+  @PostMapping
+  ResponseEntity<Void> uploadImage(MultipartFile image, String id);
+
+}