+++ ikii-back-microservices/ikii-back-commons-web/src/main/java/mx/ikii/web/commons/repository/IGenericIkiiRepository.java
@@ -1,0 +1,9 @@
+package mx.ikii.web.commons.repository;
+
+import java.util.Map;
+
+public interface IGenericIkiiRepository {
+  
+  void updateDocumentFields(String id, Map<String, Object> fields, String collectionName);
+
+}