+++ ikii-back-microservices/ikii-back-commons-web/src/main/java/mx/ikii/web/commons/util/UpdateType.java
@@ -1,0 +1,14 @@
+package mx.ikii.web.commons.util;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public enum UpdateType {
+
+  PRODUCTS("ProductBusiness"), CUSTOMER("Customer"), BUSINESS("Business");
+
+  private final String name;
+
+}